+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- `dpdocs` will now accept the eServices Membership report as an input file as an alternative to CAPWATCH.
+- Join Date and Rank Date are now a part of the Domain Model.
 
 ### Changed